feat: move CalendarButton out from base
