+++ components/base/button.tsx
@@ -1,30 +1,6 @@
-import { ButtonProps, extendVariants } from '@heroui/react';
+import { extendVariants } from '@heroui/react';
 
-import { Button as HButton, cn } from '@heroui/react';
-import COMMON_STYLES from '@/style/common';
-
-interface IButtonProps extends ButtonProps {
-  border?: boolean;
-}
-
-export default function CalendarButton({
-  children,
-  className,
-  ...props
-}: IButtonProps) {
-  return (
-    <HButton
-      className={cn(
-        'p-[10px] text-[16px] leading-[1.2] text-white rounded-[10px] gap-[10px]',
-        props.border && COMMON_STYLES.border,
-        className,
-      )}
-      {...props}
-    >
-      {children}
-    </HButton>
-  );
-}
+import { Button as HButton } from '@heroui/react';
 
 const Button = extendVariants(HButton, {
   variants: {
@@ -68,4 +44,4 @@ const Button = extendVariants(HButton, {
   },
 });
 
-export { Button, CalendarButton };
+export { Button };
